requirements.txt
beautifulsoup
django-extensions
mysql-python
requests
nltk